TEST : Upload a file while filling the form
Creates a new record, fills in the form and uploads a pdf file.
Then checks if all the data were added AND if the file was uploaded.

Tip : on Windows, place the Organ.pdf file in C:

On other systems, open the test and look for the PROTIP commentary, then
replace
the "C:\Organ.pdf" value to a suitable absolute path to the Organ.pdf
file.